Moved devices counter to the devices page in the right side.
