+++ RGBMasterUWPApp/Pages/DevicesPage.xaml
@@ -7,38 +7,79 @@
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" xmlns:muxc="using:Microsoft.UI.Xaml.Controls" xmlns:state="using:AppExecutionManager.State"
     mc:Ignorable="d"
    >
+    <Grid>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="3*" />
+            <ColumnDefinition Width="1*" />
+        </Grid.ColumnDefinitions>
+        <StackPanel Grid.Column="0">
+            <Pivot ItemsSource="{x:Bind RegisteredProviders}" x:Name="ProvidersPivot">
+                <Pivot.Title>
+                    <TextBlock Text="Devices" HorizontalAlignment="Left" />
+                </Pivot.Title>
+                <Pivot.HeaderTemplate>
+                    <DataTemplate x:DataType="state:RegisteredProvider">
+                        <StackPanel Orientation="Horizontal">
+                            <TextBlock Text="{x:Bind Provider.ProviderName}" Margin="0 0 4 0" />
+                            <TextBlock Text="(" />
+                            <TextBlock Text="{x:Bind Devices.Count}" />
+                            <TextBlock Text=")" />
+                        </StackPanel>
+                    </DataTemplate>
+                </Pivot.HeaderTemplate>
 
-        <Pivot ItemsSource="{x:Bind RegisteredProviders}" x:Name="ProvidersPivot">
-        <Pivot.Title>
-            <StackPanel Orientation="Horizontal" VerticalAlignment="Center">
-                <TextBlock Text="Devices" />
+                <Pivot.ItemTemplate>
+                    <DataTemplate x:DataType="state:RegisteredProvider">
+                        <StackPanel HorizontalAlignment="Left" Orientation="Vertical">
+                            <TextBox x:Name="ManualConnectionAddressTxtBx" PlaceholderText="Manually connect via IP address or unique ID" MaxWidth="360"/>
+                            <Button Content="Connect" x:Name="ManualConnectionButton" Click="ManualConnectionButton_Click" Margin="4 0 0 0" />
+                            <ListView ItemsSource="{x:Bind Devices}">
+                                <ListView.ItemTemplate>
+                                    <DataTemplate x:DataType="state:DiscoveredDevice">
+                                        <StackPanel Orientation="Vertical">
+                                            <CheckBox IsChecked="{x:Bind IsChecked}" Tag="{x:Bind}" Click="CheckBox_Click" Content="{x:Bind Device.DeviceName}" />
+                                        </StackPanel>
+                                    </DataTemplate>
+                                </ListView.ItemTemplate>
+                            </ListView>
+                        </StackPanel>
+                    </DataTemplate>
+                </Pivot.ItemTemplate>
+            </Pivot>
+        </StackPanel>
+        <StackPanel Grid.Column="1" Margin="0 64 0 0" Orientation="Vertical">
+            <StackPanel Margin="0 0 0 16">
+                <TextBlock FontWeight="Bold" FontSize="18" Text="Devices Overview" HorizontalAlignment="Center"/>
+                <RichTextBlockOverflow x:Name="DevicesOverviewDescription" />
+                <RichTextBlock OverflowContentTarget="{x:Bind DevicesOverviewDescription}">
+                    <Paragraph>
+                        Watch and control discovered and connected devices...
+                    </Paragraph>
+                </RichTextBlock>
+            </StackPanel>
+            <StackPanel Margin="0 0 0 16">
                 <Button Click="Refresh_Button_Clicked" ToolTipService.ToolTip="Refresh Devices" Background="Transparent">
-                    <FontIcon x:Name="OuterButton" Glyph="&#xE72C;" FontFamily="Segoe MDL2 Assets" FontSize="16" Foreground="Orange"/>
+                    <StackPanel Orientation="Horizontal">
+                        <FontIcon x:Name="OuterButton" Glyph="&#xE72C;" FontFamily="Segoe MDL2 Assets" FontSize="16" Foreground="Orange"/>
+                        <TextBlock Text="Refresh devices" Margin="4 0 0 0" />
+                    </StackPanel>
                 </Button>
             </StackPanel>
-        </Pivot.Title>
-        <Pivot.HeaderTemplate>
-            <DataTemplate x:DataType="state:RegisteredProvider">
-                <TextBlock Text="{x:Bind Provider.ProviderName}" />
-            </DataTemplate>
-        </Pivot.HeaderTemplate>
-
-        <Pivot.ItemTemplate>
-            <DataTemplate x:DataType="state:RegisteredProvider">
-                <StackPanel>
-                    <TextBox x:Name="ManualConnectionAddressTxtBx" PlaceholderText="Manually connect via IP address or unique ID"></TextBox>
-                    <Button Content="Connect" x:Name="ManualConnectionButton" Click="ManualConnectionButton_Click"></Button>
-                    <ListView ItemsSource="{x:Bind Devices}">
-                        <ListView.ItemTemplate>
-                            <DataTemplate x:DataType="state:DiscoveredDevice">
-                                <StackPanel Orientation="Vertical">
-                                    <CheckBox IsChecked="{x:Bind IsChecked}" Tag="{x:Bind}" Click="CheckBox_Click" Content="{x:Bind Device.DeviceName}" />
-                                </StackPanel>
-                            </DataTemplate>
-                        </ListView.ItemTemplate>
-                    </ListView>
-                </StackPanel>
-            </DataTemplate>
-        </Pivot.ItemTemplate>
-    </Pivot>
+            <StackPanel>
+                <TextBlock Text="Connected Devices" FontWeight="Bold" />
+                <ListView ItemsSource="{x:Bind RegisteredProviders}" SelectionMode="None" IsItemClickEnabled="False">
+                    <ListView.ItemTemplate>
+                        <DataTemplate x:DataType="state:RegisteredProvider">
+                            <StackPanel Orientation="Horizontal">
+                                <TextBlock Text="{x:Bind Provider.ProviderName}" />
+                                <TextBlock Text=":" Margin="0 0 4 0" />
+                                <TextBlock Text="{x:Bind Devices.Count}" />
+                            </StackPanel>
+                        </DataTemplate>
+                    </ListView.ItemTemplate>
+                </ListView>
+                <TextBlock x:Name="DevicesAmountTxtBlk" />
+            </StackPanel>
+        </StackPanel>
+    </Grid>
 </Page>
